Extract RTMPError to separate file
